refactor : 적절하지 못한 enums 라는 패키지 분류를 재조정
- asis : enums.*
- tobe : domain.*
- 도메인 패키지 하위로 이동시킴으로써 명확한 분류 기준에 의해 위치하도록 변경
