+++ src/main/java/nextstep/subway/favorite/domain/FavoriteExceptionType.java
@@ -1,4 +1,4 @@
-package nextstep.subway.favorite.enums;
+package nextstep.subway.favorite.domain;
 
 public enum FavoriteExceptionType {
     FAVORITE_FILED_IS_NOT_NULL("즐겨찾기를 구성하는 항목은 null이 될 수 없습니다.");
