+++ src/main/java/nextstep/subway/line/domain/Distance.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.domain;
 
-import static nextstep.subway.line.enums.LineExceptionType.DISTANCE_IS_MUST_BE_GREATER_THAN_1;
+import static nextstep.subway.line.domain.LineExceptionType.DISTANCE_IS_MUST_BE_GREATER_THAN_1;
 
 import java.util.Objects;
 import javax.persistence.Column;
