+++ src/test/java/nextstep/subway/line/domain/LineColorTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.domain;
 
-import static nextstep.subway.line.enums.LineExceptionType.CANNOT_EMPTY_LINE_COLOR;
+import static nextstep.subway.line.domain.LineExceptionType.CANNOT_EMPTY_LINE_COLOR;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
